requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
sqlparse==0.5.3
uritemplate==4.2.0
whitenoise==6.11.0
